+++ NEWS.md
@@ -20,6 +20,7 @@ This release simplifies the "key" structure. The nesting and crossing definition
 * `new_tsibble()` creates a subclass of a tsibble.
 * Renamed `fill_na()` to `fill_gaps()`, for more expressive function name and consistency to `has_gaps()` and `count_gaps()`. Soft-deprecated `fill_na()`. (#71)
 * Added `is_duplicated()`, `are_duplicated()` and `duplicates()`.
+* If `POSIXct`, time zone will be displayed in the header via `print()`.
 
 ## Bug fixes